Content with renaming for now
Ideally, we should use ./posts/SomePostName.html for the posts, or ./news/SomeNewsName.html for the news, rather than the current ./SomePostName.html URLs. That is dependent on Jekyll. Then, we need to edit index.html such that it is not the default one to open but rather opens About.md as the "./" resource.